+++ src/app.lisp
@@ -1,7 +1,9 @@
 (defpackage #:ultralisp/app
   (:use #:cl)
+  (:import-from #:flamegraph)
   (:import-from #:weblocks/app
-                #:defapp))
+                #:defapp)
+  (:import-from #:weblocks/request-handler))
 (in-package ultralisp/app)
 
 
@@ -10,3 +12,11 @@
   :description "The UltraLisp.org server."
   :autostart nil
   :debug t)
+
+
+(defmethod weblocks/request-handler:handle-request :around ((app app))
+  (let ((path (weblocks/request:get-parameter "flame")))
+    (if path
+        (flamegraph:save-flame-graph (path)
+          (call-next-method))
+        (call-next-method))))