+++ ChangeLog.rst
@@ -6,6 +6,9 @@
 ====================
 
 * Front page loading was optimized and now it should load few times faster.
+* Quicklisp was upgraded from 2019-12-27 to 2020-02-18.
+* Ultralisp was upgraded from 20200121194004 to 20200222170508.
+* Added ability to generate flamegraph data in special debug mode.
 
 0.14.12 (2020-01-27)
 ====================